Updated gptokeyb for openxcom2
Updated gptokeyb to delete extra characters in interactive text input mode, to facilitate modifying existing text (and removing all existing text) in addition to allowing text to be added and deleted. This improves functionality for openxcom2, which requires editing of soldier/aircraft names, as well as entry of new text for user name, etc.